Added to title
Objective was to change the icon in the title/tab of sharebb.
How it was done: Got the .ico file from a compressed with faviconio. Then in the head with link and rel as icon, but it can also be called shortcut icon which would be considered first in line and then with the href add the path to favicon.ico